requirements.txt
python==2.7
Django==1.7
wsgiref==0.1.2
gevent-socketio